+++ designer/client/package.json
@@ -87,6 +87,7 @@
         "remark-directive-rehype": "0.4.2",
         "remark-html": "16.0.1",
         "reselect": "4.1.5",
+        "suspend-react": "0.1.3",
         "url-join": "5.0.0",
         "use-debounce": "3.4.3",
         "uuid": "8.3.2"
