+++ designer/client/src/components/common/error-boundary/fallbackComponent/index.ts
@@ -2,3 +2,4 @@ export * from "./FullPageErrorBoundaryFallbackComponent";
 export * from "./DialogErrorFallbackComponent";
 export * from "./ToolbarErrorFallbackComponent";
 export * from "./RouteErrorFallbackComponent";
+export * from "./TextErrorBoundaryFallbackComponent";
